hwc: scale cloned UI/video to full screen with correct aspect ratio
Added support to scale the cloned UI to full screen.  Screen aspect
ratio is detected using EDID information, and UI aspect ratio is
preserved on the external display.  For now we are hardcoding square
pixels on the internal display.

In docking mode, the first NV12 layer is scaled to full screen on
the external display.

In mirroring mode, the whole UI and any video layers are scaled
to full screen.

Change-Id: Ib4887fb341c22758de7a9c2fc89a9ff660d4fbd3
Signed-off-by: Erik Gilling <konkers@android.com>