CameraHal: fix FD coordinate calculation
The android expects coordinates for 180degree rotation
to point the top left corner of the preview. When rotation
is applied the ducati is returning the faces with respect
to the camera orientation, rathar than the phone orientation.
In this case both the nTop and nLeft needs to be offset with
the size of the face, as the rotation counts for both of them.

Change-Id: I28cd8e4c0c09e6d115f24f227c8044e514a42fef
Original-author: Vladislav Hristov <vhristov@mm-sol.com>
Signed-off-by: Vladislav Hristov <vhristov@mm-sol.com>
Signed-off-by: Tyler Luu <tluu@ti.com>