omap4xxx: camera: update Focus areas in setParameters, not in setFocusMode
Fixes b/5429270

When you touch to focus, the camera app calls setParameters to pass the focus
areas to the Ducati, followed by a call to autoFocus to start AF.  Because the
focus areas were not getting updated instantly in the camera HAL before AF
began, on the one hand, and because the Ducati locks the focus regions after AF
begins, the AF region settings were being ignored.

This patch allows focus areas to be set immediately.  Region priority is set in
doAutoFocus() when non-zero area is set in FOCUS_REGION.

Change-Id: I90c86f9de0d090f24b8595b325dbf06f1f718630
Signed-off-by: Iliyan Malchev <malchev@google.com>