hwc: support CEA standard aspect ratios and finetune HDMI mode selection
Using the screen's aspect ratio does not work for CEA modes, where the
mode's aspect ratio takes precedence.

Select HDMI mode based on the following precedence:
1. pick CEA mode over VESA mode (for audio support)
2. pick a mode where we are upscaling (for scaling support)
3. pick a mode where the framebuffer fits most of the
screen (minimize black bar area).  This in effect picks a
mode with the closest aspect ratio to the frame buffer
4. pick a resolution closest to framebuffer's resolution
5. pick the highest refresh rate

Also only select an HDMI mode where we will be able to
scale the UI to the screen.

Change-Id: Ic8985436b88577cfa59f284eecbda6cdd340de6f