CameraHal: Modify exit behavior of notify thread
When the camera app dies, sometimes the release() call is coming
from the same thread as a callback instead of the main thread. When this
happens, it causes a segfault in AppCallbackNotifier because we end up
destroying AppCallbackNotifier without the NotificationThread knowing.
The subsequent waitForMsg() after release() will result in a segfault since
the message queues are deleted.

Instead of looping forever in a while loop in the notification thread,
return bool value in looper function. This gives the Thread class a chance
to exit in case of a requestExit() in the same thread. Looper function
will automatically will be called again if exit is not pending and
previous looper function call returned true.

Change-Id: I6bdc65df71e2c3ebcbd409654b828edef29b1a0e
Signed-off-by: Tyler Luu <tluu@ti.com>