CameraHal: set ZSL history len
As both ZSL and VideoSnapshot uses ZSL queue for imaga capture,
there is a common ducati parameter to configure the number of
buffers in the queue.

This patch dynamicaly sets the number of buffers in the queue,
depending on the usecase. Sets the default ZSL history len to six
in case of the ZSL mode and uses four for everything else (currently
only VideoSnapshot).

Having six buffers for ZSL mode, more RAW buffers are available in
the ducati side. This way using ZSL capture with six buffers it is
possible to compensate da preview lag. On the other side, if using
six buffers for video case, there is not enough memory, in the
ducati, for both RAW buffers pool and encoder internal memory. The
Shutter lag in the VideoSnapshot is not that visible, the time error
because of the two buffers less (max of 66ms) will not be noticed in
captured image.

Change-Id: I03bd23240158150a116f5ef15ad43c7d28ac0534
Signed-off-by: Vladislav Hristov <vhristov@mm-sol.com>