hwc: unify HDMI docking and cloning
If we have a surface with external display usage marked, switch
to docking (clone only that surface to HDMI).  Once the surface
goes away, switch back to cloning.

Also change docking mode support:

Select dockable surface based on the usage bit instead of the
color format.  For multiple such surfaces, pick the one with the
largest display area.

Change meaning of hwc_dev->ext transform bits for docking mode.
These are applied purely to the surface buffer, instead of as an
additional transform on the surface displayed on LCD.  In other
words, the surface displayed on HDMI is independent of the UI
orientation.

Change-Id: Ic6895ba0ff275ca374eb72d0fa7d93a634111d1a