CameraHal: Error handling before sending frames
Adding a few error checks before sending frames to make
sure we have the proper conditions to send a frame.

Fixes a bug where FillBufferDone gets preempted by the stop
path and by the time it returns the buffers and subscribers
are freed. Adding checks in send frame context since it is
protected by a Mutex.

Change-Id: I860bee796669606e1ec3e26e19916f386fc3e9aa