hwc: invalidate only when we have to when switching to hdmi
This moves forcing of invalidate only if all overlays are used
on LCD.  Then the layers will be first rendered with one less
overlay/half of the overlays on LCD, freeing up overlays.  On
the second invalidate the cloning will start.

This patch also resolves a potential race conditino when setting
ext_requested.

Change-Id: I0b9cf81091c9f66e37d3dab5f9f9648147c7295c