CameraHal: Check state in doAutoFocus
Check the state in doAutoFocus and return if we are not in the
AF_STATE. If user/monkey is pressing all over the screen, this
will result in multiple autoFocus/cancelAutoFocus calls until
the last autoFocus call comes when the user/monkey stop pressing
preview window. This will result in multiple CAMERA_PERFORM_AUTOFOCUS
commands being queued in OMXCameraAdapter's CommandHandler that will
all eventually need to be serviced which will result in multiple
doAutoFocus calls. This is the reason why we need to check the
state in the beginning of the function and return if we are not
focusing.

Change-Id: If245751a296604bca511ad237b769be24175cbc6
Signed-off-by: Tyler Luu <tluu@ti.com>

related-to-bug: 5472396