OMX Video Encoder : Configure Frame Rate Dynamically
This patch adds support to configure frame rate to encoder dynamically.
This is necessicated by the Variable Frame Rate support avaialble in Camera
The Frame Rate is fed to encoder periodically to maintain
the quality of recorded Video.

By default it is enabled, we can use the setprop to configire it in runtime
setprop debug.vfr.enable 0 (disable)
setprop debug.vfr.enable 1 (enable)

Ported the changes from GB(#13338) and modified it according to the Ducati
side updates

Change-Id: Id2d38c476d722c40f54a561b268fbe5f8efe7526
Signed-off-by: Prasanna Kumar M.R. <x0118059@ti.com>