CameraHal: Add 2592x1728 and 2592x1458 image capture resolutions
Part of the fix for b/5546023

Adding new resolution to support 16:9 and 3:2 aspect ratio.

Note: This allows to keep same aspect ratio between video preview
and video snapshot.

This change requires a matching Ducati update.  For the power savings to take
effect, the new resolution must actually be used by the application.

Change-Id: I87d5f3155856b0c2c83763349843628eeadfacfc
Signed-off-by: Gilles-Arnaud Bleu-Laine <gilles@ti.com>
Signed-off-by: Iliyan Malchev <malchev@google.com>