CameraHal: Fix preview not starting after surface is abdandoned
If an application gets backgrounded (HOME key) without stopping
preview and releasing camera, preview would previously not restart
when application is resumed. ANativeWindowDisplayAdapter would
detect that the preview surface was abandoned and stop sending
buffers to CameraAdapter.

We were previously not doing anything when CameraService called
setPreviewWindow when a display adapter was already created. To
fix this issue, we need to send the preview window and restart
preview.

Change-Id: I36089b0046f9861897d32c50b52128a41e2de2a5
Signed-off-by: Tyler Luu <tluu@ti.com>