CameraHAL: Fix supported 2d preview resolutions
Fix for adding too many preview resolutions.
Portrait ones need to be in the separate array.
Otherwise when we call encodeSizeCap with landscape
ones, some of the portrait are also added to supported
values. Same problem appears if we call it with
portrait resolutions. In this case some of the
preview ones are added.

Change-Id: Ifa8ede1e030efaca7d233cc015ea574c1a897380
Signed-off-by: Ivan Evlogiev <x0099089@ti.com>