CameraHAL: Fix some corner cases while handling Variable FPS Handling
Port FPS takes lower priority than VFR. But there was invalid check to take of
this priority which resulted in incorrect setting of VFR in some corner cases.
This patch makes sure Port FPS is considered only if no Range is Provided.
This patch partially resolves the DR : OMAPS00254364
This patch along with #16816 in Application completely resolves the DR.

Signed-off-by: Sathya Prakash M R <a0393677@ti.com>
Change-Id: I8aa2b5019ddd433d82fff1bad65f8cd16d2778df