[DOMX] Fix Buffer Map type during free buffer
The buffer sent from RPC calls (UseBuffer, AllocateBuffer, FreeBuffer)
will get mapped to approriate space depending on the map type that is
set in the packet.

Even though a pBuffer was sent in the packet, the map type
(RPC_OMX_MAP_INFO_TYPE) was set to RPC_OMX_MAP_INFO_NONE.
This resulted in conversion not being done in rpmsg.

This patch sets the RPC_OMX_MAP_INFO_TYPE correctly in the
packet to RPC_OMX_MAP_INFO_ONE_BUF. With this rpmsg will
convert the pBuffer address to ducati side (tiler) address.

Change-Id: I722b618687e178817e2e8ffdb9670e7a5f54d5e9
Signed-off-by: Pradeep Venkatasubbarao <pradeepv@ti.com>