Add index OMX_TI_IndexParamComponentExpectedSuspensionState
Add support to query using structure OMX_PARAM_SUSPENSIONTYPE
over the index OMX_TI_IndexParamComponentExpectedSuspensionState to
indicate the present or anticipated suspension status of OMX
camera if the IL Client tried to continue using OMX camera. If
OMX-camera is or anticipates it would be blocked on resources
if transitioned from states Loaded to Idle or beyond, it returns
the eType as OMX_Suspended and IL Clients may choose not to
continue using OMX-camera at present and try to use later on.
If the resources are available at present and it already has
or anticipates to get the resources, then the eType is returned
as OMX_NotSuspended.

Signed-off-by: Abhishek Ranka <asranka@ti.com>