Encoder: Add a custom index to retrieve the component handle
For slice based processing between camera and encoder, a tunnel must be set up
between camera and encoder. Therefore, the application requires access to the
component handle. However, the handle is hidden in lower layers of the framework.
Therefore, added a custom Index that applications can use to read the component handle.

Change-Id: I8dd24abe5d78bf188b76d48d88586cb923146679
Signed-off-by: Jorge E. Solano <jsolano@ti.com>
Signed-off-by: Anu Sundararajan <sanuradha@ti.com>

Conflicts:

	omx_core/inc/OMX_TI_Index.h

Signed-off-by: Brijesh Nekkare <b-nekkare@ti.com>