DO NOT MERGE
HWC: OMAP4: disable the first frame display reset

A hack to reset the display on the first frame was put in previously
to allow for a seamless transition between the boot logo from
the bootloader to the boot animation.

In our case, the kernel/hwc has already scanned and set the proper
hdmi mode, so we dont want to bounce the panel.

Change-Id: I35a04e1cc25504685bdb71b03cee29b9c0fce1d7