CameraHAL: Enable Sensor Overclocking for OV14825 for 60 FPS
OV14285 needs sensor overclocking to get 60 fps for
1080p resolution. This patch enables overclocking
when FPS is set to 60 and resolution is 1080p.

This patch also adds the 60fps to the capabilities.
So that when ducati gives the capability for 60fps
cameraHAL will populate this to application.

This also restarts the preview when the FPS range is
reduced from 60fps to 30fps.

Change-Id: Ib7c34a6581d3f0c9bd555a1b444b91a7cd2fda4c
Signed-off-by: Sathya Prakash M R <a0393677@ti.com>
Signed-off-by: Solaiyappan Saravanan <saravanan.s@ti.com>