hwc: added support for ext display back buffer allocation
When the ext display transform is different from the FB,
the mirroring/cloning is achived with the help of back-buffers,
which are allocated in TILER2D space to get the required
transformation. This patch adds support in HWC to detect
the ext display transform and allocate back buffers from
TILER2D space and program dsscomp accordingly.

Signed-off-by: Sunita Nadampalli <sunitan@ti.com>

hwc: make file fix for local include path

the local include path for ion is changed from base level
to local. this is required to allow compiling at hwc level.

Change-Id: Ie1e9e6688652fdc8aa496dfc367d385aca362a27
Signed-off-by: Sunita Nadampalli <sunitan@ti.com>