DO NOT MERGE:
OMAP4: HWC: Don't scale on non-1080p monitors

Instead, we're going to use WindowManager's setForcedDisplaySize()
API to have the apps draw to windows that are a subregion of the
framebuffer and only show that on the HDMI output.

We set a property when the mode is set so that Java code elsewhere
(going into TungstenLEDService in a separate CL) can read it
and pass it on to the WindowManager.

This reverts a bunch of the scaling code that was introduced
in Saket's change:

8e8bf69d94423df082460c43d818e58046b2d1de

Change-Id: Ic6b12c5552bc75aee762ee71dd786d17da1f911f
Signed-off-by: Mike J. Chen <mjchen@google.com>