DOMX: Remove stride based classification of 1D or 2D buffers
stride value (4096) was used to determine 2D buffers in

 - PROXY_UseBuffer
 - PROXY_AllocateBuffer
 - PROXY_FreeBuffer

In case of PROXY_UseBuffer - GrallocPointers, BufferDescriptorVirtual2D
are always 2D and EncoderMetadataPointers are always 1D. So the stride check
was not required as none of the handled buffer types coexist in 1D and
2D forms.

In case of PROXY_AllocateBuffer - Tiler 2D buffers are not implemented and
a request for such a buffer was detected using nStride == LINUX_PAGE_SIZE
condition. But tiler 2d buffers are presently not used and hence, this
check is removed.

In case of PROXY_FreeBuffer - BufferDescriptorVirtual2D (always 2D) is
handled within the stride check. So the stride check was not required
as it does not coexist in 1D and 2D forms.

This patch will enable usecases that require 2D buffers of multiple
stride values. Previously, the filling of pAuxBuf1 was hampered by
the stride check.

Change-Id: I31a26442dc8401f18edf5cc23ed0e7697336c72c
Signed-off-by: Yordan Lilov <ylilov@mm-sol.com>
Signed-off-by: Vidhoon Viswanathan <vidhoon@ti.com>