Camera HAL: Second stereo camera added
Supported number of cameras is increased to 4 accordingly.
The way number of cameras is detected from the OMX camera adapter is changed as
not every camera has its own camera index. Stereo cameras are distinguished by the stereo operating mode
but have the same indexes as the mono cameras. On OMX init time when capabilities are gotten
a container is filled that maps each Android camera index to the corresponding OMX camera settings -
OMX cam index and operating mode.
OMX_TI_StereoSensor index is left for backward compatibility with older Ducati images.
CameraService max number of cameras must also be increased to 4. There is another framework/base
patch that does just that.

Change-Id: I9ea9fb8454d09380612c19f6240457ab36c773e9
Signed-off-by: Pavel Nedev <pnedev@mm-sol.com>