CameraHAL: Fix for RawCapture & YuvCapture failure.
This patch fixes the issues with UseBuffer failure for
RawCapture. This is due to the CPCAM change with the buffer
handling.
pBufferHdr-pBuffer will no more hold the actual raw buffer.
pBufferHdr-pAppPrivate should be used to get the CameraBuffer
address.

This patch also fixes the yuvCapture (/data/misc/camera/YuV_PiCtUrEs/).

Change-Id: Ic4d9a25221d1916b33cd643ed810fa2a863d8cf9
Signed-off-by: Solaiyappan Saravanan <saravanan.s@ti.com>