[domx-cam]: Patch to fix buffer allocation issue
Description: Buffer allocation for internal memory in case of slice
based processing was allocated by proxy camera component. And this
buffer was always corresponding to 1080p resolution. This is modified
to allocate to the needed amount depending on the resolution request.
Also, there is a small clean up where SetParam was earlier used in
slice mode, is now removed as it is un-necessary now.

CQ:OMAPS00268023

Change-Id: Ib551f72eb7b00988944588252d79854d7d8436a1
Signed-off-by: Phanish Hanagal <phanish.hs@ti.com>