DOMX: Fix to release the shared ion buffer
In secure playback case, we dont map the ion buffers. We allocate
the buffer and perform ion share to share it across to other components.
In this case, we also need to close the fd inadditional to the existing
ion free call inorder to ensure that buffer gets released as otherwise
the reference count is still valid in ion driver and is not freed.

Change-Id: I621077e132af55ada2a5ab0c7f4a01ed7afc7d02
Signed-off-by: Prasanna Kumar M.R. <prasanna.kumarmr@ti.com>