[OMX] Propogate GetParam IndexUseBufferDescriptor to Ducati
The ducati patches:
https://gerrit.ext.ti.com/gerrit/omap/#change,8128
https://gerrit.ext.ti.com/gerrit/omap/#change,8261

implement Get/Setparam for OMX_TI_IndexUseBufferDescriptor at Ducati.
The implementation of OMX_TI_IndexUseBufferDescriptor mandates ducati side
OMX components to be aware of attributes such as buffer type to
proceed with buffer processing appropriately.

This patch propogates the GetParameter call that was so far handled
internally within DOMX to ducati OMX component utilizing the
GetParameter implementation at components. Now DOMX is  aware
of component defaults to the structure related to this index.

Change-Id: Icad3abb5236c927edda5ad668ac7ef89db92dfef
Signed-off-by: Vidhoon Viswanathan <vidhoon@ti.com>