[DOMX] SVC Encoder
This patch adds support for SVC Encoder.
It is different from the previous patchset.
Previous patchset is for svc on top of domx for ducati 2.1.16

This patchset is to support SVC Enc for ducati 2.2.
This patchset takes care of the present DOMX Common Encoder changes.

Ducati patches dependent on this patch:
https://gerrit.ext.ti.com/gerrit/omap/#q,status:open+project:repo/ducatimm+branch:dev_master+topic:SVCE,n,z

Signed-off-by: Saurabh Bipin Chandra <a0131926@ti.com>
Change-Id: I9b4033dc4e1209dfefef34637ed18f15ce1bbc92