gcbv-user: improve check for fill condition
When a usecase generates a fill with a 1x1 destination
we were hitting the blt rather than fill path. This change
improves the check for fill to validate whether there is
a virtual address which the fill code needs anyway as it
directly accessing the memory provided by the src1 buffer
descriptor.

Change-Id: I7010075d268f9dd57714119d088e7bc5c7bf17e2
Signed-off-by: Craig Stout <craig.stout@ti.com>