PROXY_BUFFER_ACCESSOR: interfaces to all types of accessors
This patch introduces a new struct PROXY_BUFFER_ACCESSOR.
This contains all possible accessors for a buffer including:

*buffer handle		pBufferHandle
*context free FD 	bufferFd
*Registered handle	pRegBufferHandle
*Mapped buffer address	pBufferMappedAddress

Three elements of this struct are included in the
PROXY_BUFFER_INFO struct replacing the following
elements that were not well organized:
		OMX_PTR pYBuffer;
		OMX_PTR pMetaDataBuffer;
		int mmap_fd;
		int mmap_fd_metadata_buff;
		OMX_PTR pRegisteredAufBux0;
		OMX_PTR pRegisteredAufBux1;
		OMX_PTR pRegisteredAufBux2;

This patch also removes the ION dependency from these elements.
In addition, all changes to introduce this new struct and
element in PROXY_BUFFER_INFO are included in this patch.

Tested on DB 285 with android and OMTB sanity.

Change-Id: I6f4ea6239bf6b8d5395fece04008f6b8298f0a95
Signed-off-by: Vidhoon Viswanathan <vidhoon@ti.com>