hwc: Use single instance of counts
After we needed to have access to layer statistics in set(), a
copy of counts was added to hwc_dev. The copies are synced at
the end of gather_layer_statistics() and if some function will
update counts after that we are going to have two copies that
are out of sync.

Change-Id: I537f645bb54e275221ff9bd8fad632eb4a1169b1
Signed-off-by: Mykola Ostrovskyy <mykola@ti.com>