hwc: rgz: Introduce concept of framebuffer state
A framebuffer state consists of a list of regionizer layers, each one has
stored a hwc layer, buffer index and a dirty counter. This represents what
the blitter has drawn in such framebuffer. For now only the current
framebuffer state is stored.

This change allows to introduce the geometry change handling in a easier
way since multiple framebuffer states are need for this purpose.

Additionally this patch addresses an inconsistency of using signed and
unsigned integers while accessing rgz_layerno field in the rgz structure
leading to warning messages that are difficult to deal with without making
explicit type casts.

Change-Id: I8bb5410b263bb734c52d445b4129a5dbea375225
Signed-off-by: Gustavo Diaz Prado <a0273371@ti.com>