hwc: rgz: Store a copy of the hwc layer data instead of pointer
Regionizer layers now store hwc layer data copies instead of just
pointers. This allows the framebuffer state to have additional
information to compare itself to another state.

Since the hwc layer pointer was used to know if a regionizer layer
was the background, with this change, now a unique buffer index was
created to distinguish it.

Change-Id: I897ee7d8e7962b7dafed7bce4db7a7e8a45ab08f
Signed-off-by: Gustavo Diaz Prado <a0273371@ti.com>