hwc: rgz: Handle geometry changes efficiently
Allow the regionizer to analyze the previous, current and target
framebuffer states to draw only the needed areas of the screen.

It is required that SurfaceFlinger supports the HWC_EXTENDED_API and
the HWC_EXTENDED_OP_LAYERDATA operation to enable this feature.

With this change the following scenarios are handled efficiently:

A) When a layer moves, only draw the damaged area between the current
and the target framebuffer state instead of redrawing the whole screen

B) When a layer appears, only draw the area of the new layer, not the
whole screen

C) When a layer disappears, only draw the area where the layer used to
be instead of the whole screen

This change is also compatible with the dirty count per layer, that is,
when a layer in the same position and has the same content for multiple
frames.

Change-Id: I281567440564fd4e9d5865a392676d18653e9ecd
Signed-off-by: Gustavo Diaz Prado <a0273371@ti.com>