[PROXY VIDDEC] Include PROXY_VIDDEC_ComponentDeInit method
This patch includes a local ComponentDeInit method for
common video decoder that would be used now extensively
with respect to GLBCE mode for playback.
Safely, for such customizations required, this function
has plugged in a call to GLUE_ComponentDeInit that can
now be used to house all customizations needed at
ComponentDeInit phase.

Change-Id: Iee2a13ec08e6943ca58e89454343d22c41056cb6
Signed-off-by: Vidhoon Viswanathan <vidhoon@ti.com>