[MPEG4E] Replicate H264E changes in MPEG4E
This patch replicates the following changes done in H264E
to optimize handling of EncoderMetaData in MPEG4E:

http://review.omapzoom.org/#/c/23320/
http://review.omapzoom.org/#/c/24029/
http://review.omapzoom.org/#/c/25497/

In addition it also an older change related to NPA mode
disable on encoder proxy and move that to ducati enc component.
This is old and not present in gerrit. Hence I have listed
the commit message and change id of this change for
future reference:

    commit dc35be139302e74411c930796b3bbed3321f6509
    Author: Sarthak Aggarwal <sarthak@ti.com>
    Date:   Sat Aug 27 11:45:59 2011 +0530
    Enable NPA mode on h264 encoder.

    This patch completes the i/p buffer optimization in case of metadata buffers
    The patch also sets the correct alloc length and filled length to ducati side OMX encoder, while gives the metadata size to the OMX IL client
    change id of org change: I0ac758c367ceb9affa6e47fedbc20a80d7058b31

Change-Id: Ib23a1afd5e02c01547b1224bdc164c8293b3bf9e
Signed-off-by: Vidhoon Viswanathan <vidhoon@ti.com>