Update patch set 6
Patch Set 6:

I would say merge the "camera: Fix memory leak in stop recording usecase." commit. Maybe submit it to gerrit, create a topic, and put that commit and this commit in the same topic, so we know they're linked. Or specify it in the commit message.

Patch-set: 6
Label: Code-Review=0