omaplfb: add back support for vram-based flip chain buffers
Based on the older, in-kernel omaplfb code, which we used on DDK 1.9.
The purpose of this is to save a bit of the tiler address space, as
without this change, the secondary display buffers end up allocating on
tiler2d.

Tested by switching back the main display buffers to vram.

Change-Id: I525d4ff46cf79e74ac54e9275b3d1cb2a0d65321
Signed-off-by: D. Andrei Măceș <andrei@unlegacy-android.org>