OMX_Core: Fix Leak in Case of Init Failure
In the case where the component initialization fails,
pModule was not being set back to NULL, leading to that
instance being left marked as "in-use". After enough
initialization failures, no more instances could be
created due to the array being full.

This patch fixes the issue by setting the pModule[i] to
NULL in case of initialization failure.

Change-Id: I66830ebb4b07c96a193ea4b7e6fcc18b37e0d0ca
Signed-off-by: Angela Stegmaier <angelabaker@ti.com>
Signed-off-by: D. Andrei Măceș <andrei@unlegacy-android.org>