memplugin: fix bufferFd data type
File descriptors should be stored in a regular signed int.
Otherwise, if we check the fd validity (>= 0), it'll always equal to
true, even if we previously invalidated it (set to -1).

Also, check the fd validity in MemPlugin_ION_Free as well.
In some cases, we want to unmap a buffer, but we don't have a valid
associated fd (for example, in camera proxy).

Change-Id: I669b983a4b179b6a2899c8bed2360e9a1976b688
Signed-off-by: D. Andrei Măceș <andrei@unlegacy-android.org>